Added -k argument for key. Have not tested yet.

+++ airpg/scripts/airpg_analyze.py
@@ -95,9 +95,10 @@ def main(args):
     mail = args.mail
     query = args.query
     iro = ir_operations.IROperations(log)
-    EI = entrez_interaction.EntrezInteraction(log)
+    EI = entrez_interaction.EntrezInteraction(args.mail, args.api_key, log)
 
-  # STEP 2. Read in accession numbers to loop over
+
+# STEP 2. Read in accession numbers to loop over
     tio = table_io.TableIO(args.infn, args.outfn, args.blocklist, logger = log)
     tio.remove_blocklisted_entries()
 
@@ -128,21 +129,13 @@ def main(args):
 
         # Step 3.1. Get flatfile
         if not os.path.isfile(os.path.join(args.recordsdir, accession + ".tar.gz")):
-            log.info("Saving GenBank flat file for accession `%s`." % (str(accession)))
-            if EI.internet_on():  # Check if internet connection active
-                try:
-                    fp_entry = EI.fetch_gb_entry(accession, acc_folder)
-                except:
-                    log.warning("Error retrieving accession `%s`. Skipping this accession." % (str(accession)))
-                    shutil.rmtree(acc_folder)
-                    continue
-            else:  # If no internet connection, raise error
-                raise Exception("ERROR: No internet connection.")
-        else:
-            log.info("GenBank flat file for accession `%s` already exists. Extracting existing file." % (str(accession)))
-            tar = tarfile.open(os.path.join(args.recordsdir, accession + ".tar.gz"), "r:gz")
-            tar.extractall(acc_folder)
-            tar.close()
+            fp_entry = prefetched.get(accession)
+            if not fp_entry:
+                log.warning("No prefetched file for accession `%s`. Skipping." % accession)
+                shutil.rmtree(acc_folder)
+                continue
+            # Copy from recordsdir to acc_folder
+            shutil.copy(fp_entry, acc_folder)
             fp_entry = os.path.join(acc_folder, accession + ".gb")
 
         # Step 3.2. Parse and analyze flatfile
@@ -231,6 +224,7 @@ if __name__ == "__main__":
     parser.add_argument("--recordsdir", "-r", type=str, required=False, default="./records/", help="(Optional) Path to records directory")
     parser.add_argument("--datadir", "-d", type=str, required=False, default="./data/", help="(Optional) Path to data directory")
     parser.add_argument("--verbose", "-v", action="store_true", required=False, default=False, help="(Optional) Enable verbose logging")
+    parser.add_argument("--api_key", "-k", type=str, required=False, help="(Optional) NCBI API key to increase rate limit")
     args = parser.parse_args()
     #if bool(args.query) ^ bool(args.mail):
     #    parser.error("--query and --mail must be given together")
